refactor: Rename `…/package/compile/events/api-gateway/lib/usagePlan.js`
