+++ lib/plugins/aws/package/compile/events/api-gateway/index.js
@@ -9,7 +9,7 @@ const validate = require('./lib/validate');
 const compileRestApi = require('./lib/rest-api');
 const compileRequestValidators = require('./lib/request-validator');
 const compileApiKeys = require('./lib/api-keys');
-const compileUsagePlan = require('./lib/usagePlan');
+const compileUsagePlan = require('./lib/usage-plan');
 const compileUsagePlanKeys = require('./lib/usagePlanKeys');
 const compileResources = require('./lib/resources');
 const compileCors = require('./lib/cors');
